requirements.txt
streamlit==1.32.0
beautifulsoup4==4.12.3
requests==2.31.0
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy==1.26.4
langchain==0.1.11
langchain-text-splitters==0.0.1
openai==1.13.3
lxml==5.1.0
html5lib==1.1
